moved api types and formatAnimeInformation
